[#7] Fix rendering of color

+++ ofws_rendering_glium/src/vertex.rs
@@ -3,12 +3,12 @@ use ofws_core::interface::rendering::Point;
 #[derive(Copy, Clone)]
 pub struct ColoredVertex {
     pub position: Point,
-    pub color: [u8; 3],
+    pub color: [f32; 3],
 }
 
 #[derive(Copy, Clone)]
 pub struct TexturedVertex {
     pub position: Point,
-    pub color: [u8; 3],
+    pub color: [f32; 3],
     pub tc: (f32, f32),
 }
